Script 'SLoop', 'SOldLoop'. Cambios: 1 Funcionales (-) + 5 Sintácticos (+).
	Coloco en 'SOldLoop' la parte antigua de 'SLoop' que usa 'fwp_old_daq' en vez de 'fwp_daq'.

Cambios funcionales (-):
	ahora trabajo fuera de 'with' con las clases 'DAQ' y 'Task'.

Cambios sintácticos (-):
	elimino secciones que utilizaban clases 'OldTask' y 'OldDAQ';
	renombro y reordeno secciones;
	agrego comentarios.

Cambios sintácticos (+):
	muevo 'idea_ins', 'SOldLoop', 'SWriter' y 'fwp_old_daq' a nueva carpeta 'OldCode';
	renombro 'idea_ins' a 'SInsIdea'.
