+++ OldCode/SInsIdea.py
