+++ OldCode/SWriter.py
